Draw all components of glyph in the grid view
This handles recursion but does not check for loops, except insofar
as the recursion is actually recursive and the app will crash when
it runs out of stack space.

A complicated thing about components is that they can also have
recursive transformations; one component can include another component,
with a particular transformation, and then that whole entity can be included
as a component with a different transformation.