Add ScrollZoom widget, figure out layout geometry
This moves to using a real Scroll widget for our scrolling,
and adds a ScrollZoom widget to wrap it and add zoom support.

This means that scroll/zoom logic is no longer in ViewPort,
but ViewPort still exists and is used for mapping between
the view and the design coordinate spaces.