Multi-format copy
This enables the 'copy' (cmd+C) menu item, which will copy the current
glyph to the clipboard in multiple formats: glyphs.app, pdf, svg,
and as a 'code string', of piet drawing code.

This doesn't implement paste; it also doesn't respect selections,
and always copies the entirety of the current glyph.