Implement basic paste
This is a semi-accidental patch. It implements pasting data from
binary plists with the Glyphs.app style.

It does not actually work with stuff that is put on the clipboard
_by_ glyphs, because glyphs uses string plists, which the rust
plist library does not know how to parse.

However, it _does_ mean that you can copy and paste within the app,
so I think it's worth a commit just for that case. We can deal with
other formats at a later date.