Implement pasting of pdf data
This gets PDF paste working. This was a bit of a headache, because
there's an odd behaviour on macOS where it seems that paths that
are both filled and stroked are included twice, the second time
after a matrix transformation. I'm sure there's a rationale for this,
but I couldn't find it, and so we apply these transforms as we
encounter them and then remove duplicate paths at the end.