Make BezCache play nice with Data
I'm... not totally sure we need this, or will use it? But I want
to have a better way of indicating when the cache has changed,
for updating the grid view during editing. That said, I should maybe
delete all this and figure out a better way to handle that.