Remove all caching
This removes all of the fancy caching stuff we were doing for outlines,
and attempts to generally simplify how we represent state.

This is all motivated by a general desire to play nicely with Data.

The basic issue is that we want to make sure that when a glyph
changes that is a component of other glyphs, those changes
are seen in each grid view cell as well as in each open window
that is editing a glyph that uses the modified glyph as a component.

Caching was making this trickier, as were the various different
ways we represent a glyph; the native UFO representation, our internal
representation, and then as a BezPath.

This simplifies that; the only representation that we care about is
the UFO one.

This isn't 100% pretty. In particular we've been hacking and slashing
in data.rs quite a bit, and some of the types no longer make total sense;
why does GlyphSet exist? But i'd like to look at that as follow-up
work.