Add Maybe widget
This is an experiment in a widget for representing `Option<T>`.
It has two branches, one with data `T` and one with data `()`, and
it internally switches betweem them as the data changes.

This is a specialization of the more interesting idea of a widget
for representing arbitrary enums, but I like the relative simplicity
of the implementation, and could imagine it being used more widely.

This uses the new widget to toggle whether or not to display a glyph
in the sidebar, based on whether a glyph is selected.