Calculate glyph codepoints based on name
This uses the aglfn list to map from names to codepoints, and
if a name does not exist it just uses the first character in
the name. Which is not correct! totally not okay! But also maybe
the least bad thing to do, for now.

This also reworks how we actually do renaming; previously we were
doing some very fancy lens stuff, but I think it's easier to just
use a command for this.