Fix EditableLabel
The previous implementation did not survive the changes to the
druid TextBox widget.

This implementation is also a bit hacky, but overall I like it
better; instead of holding onto both a label and a textbox, we
just use a textbox but synthesize its data while editing.

When we draw it as a 'label', we just grab the inner layout
object and draw it directly.