More principled API & use for EntityId
We were being a bit cavalier previously with how we manipulated
and compared EntityIds; this tries to add actual API in places
where we were previously doing a bunch of reaching into structs
and comparing things directly.

At some point we may want a more expressive representation for
an entity identifier, and this will let us be more robust
in that case.