Make saving more atomic
This adds an additional step while saving.

previously we would first move the existing file to a backups
directory, and then write the new data in place. With this change,
we now first try to write the data to a temporary path, and then
on success we try to back up the existing data before finally
moving the newly saved data to the target path.

- closes #178