Add druid-theme-loader crate
This adds the ability to load theme information from a file,
and to dynamically update the UI when that file is edited.

The new code is in a new crate, druid-theme-loader; that is
currently included in this repo.

This moves the definition of the currently used theme values
to a new file, `default.theme`. If we want to add more values to
this file, it is a two part process:

- define the key in the theme.rs file, and add it to the list in
the loadable_theme! macro
- add the key and value to the resources/default.theme file

Currently the theme supports floats and colors. We can add more
types as needed over time.

Note: the current implementation requires a new ThemeLoader for
each window, which is not ideal; it means each time the theme changes
we will be loading and parsing the file n times, in n separate threads.
This is okay as a proof of concept, but not ideal longer-term.