Move Entity, PathPoint, PointType to point.rs
This is preliminary work towards spline integration; I would like
to start breaking up path.rs in order to share as much code as
is reasonable between the bezier and hyperbezier splines.

In particular I want to use the same type to represent the actual
vec of offcurve and on-curve points; there's a bunch of fiddly logic
here, and I think it can be reasonably well isolated and reused.

This doesn't do that (although it *does* change the representation
of PointType, in anticipation of having auto points) but it lays
the groundwork.