More fixes for point deletion
first: If we delete the first or last point in an open path,
and that point is part of a curve segment, we should
also delete the segment's off-curve points.

second (a subtle one): by computing the points to delete all at
once, up front, we allowed invalid states that were dependent on
the order of deletions; the fix for that is to delete the
points iteratively.