Remove the Tools menu bar item for now.
On the GTK platform,   there is no straightforward way to disable hot­
keys while editing text,  this means that the menu bar will always re­
ceive these keys before the text input does because AccelGroups always
receive keys before DrawArea. Furthermore,  because  menus are global,
there is nothing available to  menus::make_menu  to disable these menu
items when the focused window is not a glyph editor window,   and this
will be a problem again down the line, when there are non-numeric text
inputs  (e.g. a search box for selecting components to add to a glyph)
in the glyph editor window.

In the future, we can add hax to work around the GTK behavior, or may­
be we'll be onto bigger and better things re the X11 and Wayland plat­
forms by then.

In the mean time,   the hotkeys still work in the glyph editor window,
where they actually make sense;  maybe tooltips should be used to make
them discoverable, or maybe we just draw hotkeys on/in the tool icons.